Add -Wdouble-promotion switch to clang/gcc compilers
It introduces a lot of ugly casts. Long term the floats should probably
be changed to doubles everywhere since we're only compiling for 64 bit
x86.

+++ src/overlaycontroller.cpp
@@ -1076,7 +1076,7 @@ void OverlayController::playAlarm01Sound( bool loop )
 
 void OverlayController::setAlarm01SoundVolume( float vol )
 {
-    m_alarm01SoundEffect.setVolume( vol );
+    m_alarm01SoundEffect.setVolume( static_cast<double>( vol ) );
 }
 
 void OverlayController::cancelAlarm01Sound()
